Use spring boot managed rest template (#20)
Removed @AutoConfigureWebClient(registerRestTemplate = true) since it’s equivalent to manually declaring a RestTemplate built from RestTemplateBuilder. 
This is no longer needed because services now build their own RestTemplate from the springboot RestTemplateBuilder according to the modern best practices (Requiring a restTemplate bean was the old springboot 1.x technique before the introduction of RestTemplateBuilder). Note that in the test RestClientTest already provides the builder because it implicitely does @AutoConfigureWebClient like in production or in @SpringBootTest

Signed-off-by: benrejebmoh <mohamed.ben-rejeb@rte-france.com>
Co-authored-by: Jon HARPER <jon.harper@rte-france.com>